+++ src/examples/dotenv/mod.rs
@@ -30,7 +30,7 @@ PORT=8080
 // Cargo.toml
 
 pub fn dotenv_cargo_toml(name: &str, fluxor_version:  &str) -> String {
-    let crate_name = to_crate_name(name);
+    let package_name = to_crate_name(name);
     format!(
             r#"[package]
 name = "{}"
@@ -40,7 +40,7 @@ edition = "2024"
 [dependencies]
 fluxor = "{}"
 "#,
-            crate_name, fluxor_version
+            package_name, fluxor_version
         )
 }